requirements.txt
pandas
scikit-learn
azureml-sdk
azure-ai-ml
kfp
mlflow
kfp-server-api
tensorflow==2.9.0

